Refine EEPROM types / flags (#17772)
