+++ Marlin/src/HAL/DUE/inc/Conditionals_post.h
@@ -22,7 +22,7 @@
 #pragma once
 
 #if USE_FALLBACK_EEPROM
-  #undef SRAM_EEPROM_EMULATION
-  #undef SDCARD_EEPROM_EMULATION
   #define FLASH_EEPROM_EMULATION
+#elif EITHER(I2C_EEPROM, SPI_EEPROM)
+  #define USE_SHARED_EEPROM 1
 #endif
