+++ Marlin/src/HAL/ESP32/HAL.h
@@ -109,12 +109,6 @@ int freeMemory();
 
 void analogWrite(pin_t pin, int value);
 
-// EEPROM
-void eeprom_write_byte(uint8_t *pos, unsigned char value);
-uint8_t eeprom_read_byte(uint8_t *pos);
-void eeprom_read_block (void *__dst, const void *__src, size_t __n);
-void eeprom_update_block (const void *__src, void *__dst, size_t __n);
-
 // ADC
 #define HAL_ANALOG_SELECT(pin)
