+++ Marlin/src/HAL/LPC1768/eeprom_flash.cpp
@@ -58,6 +58,8 @@ static uint8_t ram_eeprom[EEPROM_SIZE] __attribute__((aligned(4))) = {0};
 static bool eeprom_dirty = false;
 static int current_slot = 0;
 
+size_t PersistentStore::capacity() { return EEPROM_SIZE; }
+
 bool PersistentStore::access_start() {
   uint32_t first_nblank_loc, first_nblank_val;
   IAP_STATUS_CODE status;
@@ -121,8 +123,6 @@ bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t 
   pos += size;
   return false;  // return true for any error
 }
-
-size_t PersistentStore::capacity() { return EEPROM_SIZE; }
 
 #endif // FLASH_EEPROM_EMULATION
 #endif // TARGET_LPC1768
