+++ Marlin/src/HAL/SAMD51/HAL.h
@@ -114,12 +114,6 @@ void HAL_clear_reset_source();  // clear reset reason
 uint8_t HAL_get_reset_source(); // get reset reason
 
 //
-// EEPROM
-//
-void eeprom_write_byte(uint8_t *pos, unsigned char value);
-uint8_t eeprom_read_byte(uint8_t *pos);
-
-//
 // ADC
 //
 extern uint16_t HAL_adc_result;     // Most recent ADC conversion
