+++ Marlin/src/HAL/SAMD51/inc/Conditionals_post.h
@@ -23,4 +23,6 @@
 
 #if USE_FALLBACK_EEPROM
   #define FLASH_EEPROM_EMULATION
+#elif EITHER(I2C_EEPROM, SPI_EEPROM)
+  #define USE_SHARED_EEPROM 1
 #endif
