+++ Marlin/src/HAL/STM32/HAL.h
@@ -192,16 +192,6 @@ static inline int freeMemory() {
 #pragma GCC diagnostic pop
 
 //
-// EEPROM
-//
-
-// Wire library should work for i2c EEPROMs
-void eeprom_write_byte(uint8_t *pos, unsigned char value);
-uint8_t eeprom_read_byte(uint8_t *pos);
-void eeprom_read_block(void *__dst, const void *__src, size_t __n);
-void eeprom_update_block(const void *__src, void *__dst, size_t __n);
-
-//
 // ADC
 //
