+++ Marlin/src/HAL/STM32/eeprom_sram.cpp
@@ -2,7 +2,9 @@
  * Marlin 3D Printer Firmware
  *
  * Copyright (c) 2020 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
- * SAMD51 HAL developed by Giuliano Zaro (AKA GMagician)
+ * Copyright (c) 2016 Bob Cousins bobcousins42@googlemail.com
+ * Copyright (c) 2015-2016 Nico Tonnhofer wurstnase.reprap@gmail.com
+ * Copyright (c) 2016 Victor Perez victor_pv@hotmail.com
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,49 +20,45 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
  */
-
-#ifdef __SAMD51__
+#if defined(ARDUINO_ARCH_STM32) && !defined(STM32GENERIC)
 
 #include "../../inc/MarlinConfig.h"
 
-#if ENABLED(EEPROM_SETTINGS) && NONE(QSPI_EEPROM, FLASH_EEPROM_EMULATION)
+#if ENABLED(SRAM_EEPROM_EMULATION)
 
+#include "../shared/eeprom_if.h"
 #include "../shared/eeprom_api.h"
 
-size_t PersistentStore::capacity() { return E2END + 1; }
-
-bool PersistentStore::access_start() { return true; }
+size_t PersistentStore::capacity()    { return 4096; } // 4K of SRAM
+bool PersistentStore::access_start()  { return true; }
 bool PersistentStore::access_finish() { return true; }
 
 bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, uint16_t *crc) {
   while (size--) {
-    const uint8_t v = *value;
-    uint8_t * const p = (uint8_t * const)pos;
-    if (v != eeprom_read_byte(p)) {
-      eeprom_write_byte(p, v);
-      delay(2);
-      if (eeprom_read_byte(p) != v) {
-        SERIAL_ECHO_MSG(STR_ERR_EEPROM_WRITE);
-        return true;
-      }
-    }
+    uint8_t v = *value;
+
+    // Save to Backup SRAM
+    *(__IO uint8_t *)(BKPSRAM_BASE + (uint8_t * const)pos) = v;
+
     crc16(crc, &v, 1);
     pos++;
     value++;
-  }
+  };
+
   return false;
 }
 
 bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t *crc, const bool writing/*=true*/) {
-  while (size--) {
-    uint8_t c = eeprom_read_byte((uint8_t*)pos);
+  do {
+    // Read from either external EEPROM, program flash or Backup SRAM
+    const uint8_t c = ( *(__IO uint8_t *)(BKPSRAM_BASE + ((uint8_t*)pos)) );
     if (writing) *value = c;
     crc16(crc, &c, 1);
     pos++;
     value++;
-  }
+  } while (--size);
   return false;
 }
 
-#endif // EEPROM_SETTINGS && !(QSPI_EEPROM || FLASH_EEPROM_EMULATION)
-#endif // __SAMD51__
+#endif // SRAM_EEPROM_EMULATION
+#endif // ARDUINO_ARCH_STM32 && !STM32GENERIC
