+++ Marlin/src/HAL/STM32F1/eeprom_wired.cpp
@@ -23,7 +23,10 @@
 
 #if USE_WIRED_EEPROM
 
+#include "../shared/eeprom_if.h"
 #include "../shared/eeprom_api.h"
+
+size_t PersistentStore::capacity()    { return E2END + 1; }
 
 bool PersistentStore::access_start() {
   #if ENABLED(SPI_EEPROM)
@@ -69,8 +72,6 @@ bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t 
   } while (--size);
   return false;
 }
-
-size_t PersistentStore::capacity() { return E2END + 1; }
 
 #endif // USE_WIRED_EEPROM
 #endif // __STM32F1__
