+++ Marlin/src/HAL/TEENSY35_36/eeprom.cpp
@@ -29,7 +29,8 @@
 #include "../shared/eeprom_api.h"
 #include <avr/eeprom.h>
 
-bool PersistentStore::access_start() { return true; }
+size_t PersistentStore::capacity()    { return E2END + 1; }
+bool PersistentStore::access_start()  { return true; }
 bool PersistentStore::access_finish() { return true; }
 
 bool PersistentStore::write_data(int &pos, const uint8_t *value, size_t size, uint16_t *crc) {
@@ -62,8 +63,6 @@ bool PersistentStore::read_data(int &pos, uint8_t* value, size_t size, uint16_t 
   } while (--size);
   return false;
 }
-
-size_t PersistentStore::capacity() { return E2END + 1; }
 
 #endif // USE_WIRED_EEPROM
 #endif // __MK64FX512__ || __MK66FX1M0__
