+++ Marlin/src/HAL/shared/eeprom_if_spi.cpp
@@ -21,15 +21,16 @@
  */
 
 /**
- * Description: functions for SPI connected external EEPROM.
- * Not platform dependent.
+ * Platform-independent Arduino functions for SPI EEPROM.
+ * Enable USE_SHARED_EEPROM if not supplied by the framework.
  */
 
 #include "../../inc/MarlinConfig.h"
 
-#if ENABLED(SPI_EEPROM)
+#if BOTH(USE_SHARED_EEPROM, SPI_EEPROM)
 
 #include "../HAL.h"
+#include "eeprom_if.h"
 
 #define CMD_WREN  6   // WREN
 #define CMD_READ  2   // WRITE
@@ -119,4 +120,4 @@ void eeprom_update_block(const void* src, void* eeprom_address, size_t n) {
   delay(EEPROM_WRITE_DELAY);   // wait for page write to complete
 }
 
-#endif // SPI_EEPROM
+#endif // USE_SHARED_EEPROM && I2C_EEPROM
