+++ Marlin/src/inc/Conditionals_adv.h
@@ -337,3 +337,8 @@
 #else
   #define SD_CONNECTION_IS(...) 0
 #endif
+
+// Flag if an EEPROM type is pre-selected
+#if ENABLED(EEPROM_SETTINGS) && NONE(I2C_EEPROM, SPI_EEPROM, QSPI_EEPROM, FLASH_EEPROM_EMULATION, SRAM_EEPROM_EMULATION, SDCARD_EEPROM_EMULATION)
+  #define NO_EEPROM_SELECTED 1
+#endif
