+++ Marlin/src/pins/stm32f1/pins_JGAURORA_A5S_A1.h
@@ -44,9 +44,11 @@
 
 //#define I2C_EEPROM                              // AT24C64
 //#define E2END 0x7FFFUL                          // 64KB
+
 //#define FLASH_EEPROM_EMULATION
 //#define E2END 0xFFFUL                           // 4KB
 //#define E2END (EEPROM_START_ADDRESS + (EEPROM_PAGE_SIZE) * 2UL - 1UL)
+
 //#define EEPROM_CHITCHAT
 //#define DEBUG_EEPROM_READWRITE
