+++ Marlin/src/pins/stm32f1/pins_LONGER3D_LK.h
@@ -147,8 +147,10 @@
 // Persistent Storage
 // If no option is selected below the SD Card will be used
 //
-//#define SPI_EEPROM
-#define FLASH_EEPROM_EMULATION
+#if NO_EEPROM_SELECTED
+  //#define SPI_EEPROM
+  #define FLASH_EEPROM_EMULATION
+#endif
 
 #undef E2END
 #if ENABLED(SPI_EEPROM)
