+++ Marlin/src/pins/stm32f1/pins_MKS_ROBIN_MINI.h
@@ -38,11 +38,13 @@
 //
 #define DISABLE_DEBUG
 
-#define FLASH_EEPROM_EMULATION
-// 2K in a AT24C16N
-#define EEPROM_PAGE_SIZE     (0x800U) // 2KB
-#define EEPROM_START_ADDRESS (0x8000000UL + (STM32_FLASH_SIZE) * 1024UL - (EEPROM_PAGE_SIZE) * 2UL)
-#define E2END (EEPROM_PAGE_SIZE - 1)
+#if EITHER(NO_EEPROM_SELECTED, FLASH_EEPROM_EMULATION)
+  #define FLASH_EEPROM_EMULATION
+  // 2K in a AT24C16N
+  #define EEPROM_PAGE_SIZE     (0x800U) // 2KB
+  #define EEPROM_START_ADDRESS (0x8000000UL + (STM32_FLASH_SIZE) * 1024UL - (EEPROM_PAGE_SIZE) * 2UL)
+  #define E2END (EEPROM_PAGE_SIZE - 1)
+#endif
 
 //
 // Note: MKS Robin mini board is using SPI2 interface.
