+++ Marlin/src/pins/stm32f4/pins_BTT_GTR_V1_0.h
@@ -32,9 +32,11 @@
 #define BOARD_INFO_NAME "BIGTREE GTR 1.0"
 
 // Use one of these or SDCard-based Emulation will be used
-//#define I2C_EEPROM
-//#define SRAM_EEPROM_EMULATION                   // Use BackSRAM-based EEPROM emulation
-//#define FLASH_EEPROM_EMULATION                  // Use Flash-based EEPROM emulation
+#if NO_EEPROM_SELECTED
+  //#define I2C_EEPROM
+  //#define SRAM_EEPROM_EMULATION                 // Use BackSRAM-based EEPROM emulation
+  //#define FLASH_EEPROM_EMULATION                // Use Flash-based EEPROM emulation
+#endif
 
 #define TP                                        // Enable to define servo and probe pins
 
@@ -389,3 +391,5 @@
   //#define DOGLCD_MOSI                     PB15
 
 #endif // HAS_SPI_LCD
+
+#undef TP
