+++ Marlin/src/pins/stm32f4/pins_FYSETC_S6.h
@@ -34,19 +34,21 @@
   #define DEFAULT_MACHINE_NAME BOARD_INFO_NAME
 #endif
 
-// change the prio to 3 , 2 is for software serial
+// Change the priority to 3. Priority 2 is for software serial.
 //#define TEMP_TIMER_IRQ_PRIO 3
 
 //
 // EEPROM Emulation
 //
-#define FLASH_EEPROM_EMULATION
+#if NO_EEPROM_SELECTED
+  #define FLASH_EEPROM_EMULATION
+  //#define SRAM_EEPROM_EMULATION
+  //#define I2C_EEPROM
+#endif
+
 #if ENABLED(FLASH_EEPROM_EMULATION)
   #define FLASH_EEPROM_LEVELING
-#endif
-//#define SRAM_EEPROM_EMULATION
-//#define I2C_EEPROM
-#ifdef I2C_EEPROM
+#elif ENABLED(I2C_EEPROM)
   #undef E2END                                    // Defined in Arduino Core STM32 to be used with EEPROM emulation. This board uses a real EEPROM.
   #define E2END 0xFFF                             // 4KB
 #endif
