+++ Marlin/src/pins/stm32f7/pins_REMRAM_V1.h
@@ -28,7 +28,9 @@
 #define BOARD_INFO_NAME      "RemRam v1"
 #define DEFAULT_MACHINE_NAME "RemRam"
 
-#define SRAM_EEPROM_EMULATION                     // Emulate the EEPROM using Backup SRAM
+#if NO_EEPROM_SELECTED
+  #define SRAM_EEPROM_EMULATION                   // Emulate the EEPROM using Backup SRAM
+#endif
 
 #if HOTENDS > 1 || E_STEPPERS > 1
   #error "RemRam supports only one hotend / E-stepper."
